requirements.txt
pandas
numpy
scikit-learn
joblib
Flask
Flask-Login
Flask-WTF
itsdangerous
Jinja2
MarkupSafe
Werkzeug
WTForms
torch
torchvision
Pillow
requests
blinker
click
colorama
